Support font cascade list.
Settings including 'font_face' and 'label_font_face'
now support customizable font cascade list.

Font cascade list are font names separated by the
comma ',' (may contain some whitespace characters).
The first font name in the cascade list will be treated
as the default font name, the rest of them are fallback
fonts (the same order as in the cascade list).

If font cascade list is not required, a single font name
will be fairly enough, just the same as before.

The order of fonts in the cascade list DOES matter.
Only characters which are not included in any particular
font will fallback to the font followed by.
Eventually, characters which are not included in any
font in the cascade list will fallback to the system
default (i.e. 'Lucida Grande').