fix crash when enabled inline_candidate
the complex logic handles the happy path, but the indices can be wrong
eg. when the preedit is a complex structure containing prompt text.

code that removes remaining code from the candidate preview is
particularly dangerous.

the current fix is not perfect and can produce wrong selection range.
let's improve it by adding librime APIs that provide segmented and
detailed data of the composition.